feat(proposals): rename the graph and handler to get governor address
